Add venv parameter to be able to specify virtual environment
By default ansible run each task on a target host by the discovered
python executable. In such case we need to install the reportportal
plugin requirements to the system level python library even if the
infrared is executed in the virtual environment with all the required
packages. To avoid that we can specify python interpreter manually.

Change-Id: I8d84ee142a70c9eb4ee7af7e6c4c49221345216e