Make reportportal plugin to generate deployment results
In order to keep the deployment information attached to the launch
we need to collect stages statuses together with some short logs.
It can be done using Jenkins pipeline REST API [1].

Collected information is saved in the XML file. It will be uploaded
to the launch as a separate test suite. Each stage will be represented
by a single test case.

To make sure that stages are listed in the correct order - timestamp
and stage duration should be accurate.

[1] https://github.com/jenkinsci/pipeline-stage-view-plugin/

Change-Id: Id7718f4b362116e95b7d00fcbd86450aa28cd100