Support alternative launch name
On OSASINFRA DFG we are using a single ocp_testing job that, depending
on the provided configuration, will run different testsuites. In order
to keep the control of the test results with ReportPortal, we need to
have different launch names depending on the configuration so we can
compare the results between equivalent runs.

This patch includes the possibility to define a new parameter on the
infrared reportportal plugin called '--launch-altname'. If the
parameter is not included, the behaviour of the plugin is the same
as before this patch.