Add an option to not upload data of passed tests
In most cases, users don't check/look at passed tests, but only skipped
or failed tests. Because of that, it may be useful to not upload the
data (system-out) of passed tests in order to save storage and
processing resources.
A new Boolean flag 'no-log-passed-tests' (default to 'false') was added
in order to control that behavior.

RHOSINFRA-4387

Change-Id: I7570233224b174f5f034d2d89e365fdcea45ed1f