Don't generate deployment XUnit by default
Eeach time that the ReportPortal InfraRed Plugin is executed with
'--import true', it creates a XUnit of the Jenkins stages ot the given
Jenkins Build URL. Besides the fact that it increases the execution
time a lot, the deployment result file that was created doesn't always
uploaded successfully (it doesn't fail!) and because of that, other
test results (test suites) from other resources are not updated in the
newly created launch as well.
This change remove the automatic creatin of this deployment XUnit file
by removing its default value in the plugin.spec and by adding a when
condition to the Ansible task that creates this file.

Change-Id: I2ba236e7e231a0416c26b73e302d1440957e8620