Adds launch validation
This change added the option to validate the newly created launch using
the '--validate-launch' flag.
ATM, it only validates the amount of tests that were uplaoded compare to
the amount of tests in the result files.

This change also added some usefult data to the return json like info
about the launch, the calculated tests and if the launch was validated
or not.

Change-Id: I690b6c2e4f46665db6253785f6f45961cb5e448b