Fix handling of empty test suites
This changes fixed how reportportal_api.py used to handle empty test
suites. In such case, the code returned a None object inside a list and
tried to create a test case from this None object.
The fix creates an empty list instead of a list with a None object and
make sure to create test items only if the list isn't empty.

Change-Id: I12512e6755a8534c0d2cfd555132eda028f1e69c