Include failure when @message is defined but empty
On OSASINFRA DFG, we are uploading JUNIT files that includes the message
attribute under the failure tag but empty. Due to that condition,
reportportal_api.py is uploading an empty string.

This change will check not only if @message exists, but also if it not
not empty.

This change is needed to use the PatternAnalysis feature of
ReportPortal. The ERROR log trace will be parsed to analyze the errors.

Change-Id: I93271412716df96004204e72b031fba155915214