Ensure launch_id after start_launch and remove SSL hack
1.
It is necessary to incorporate an explicit check after
invoking start_launch to ensure that the launch_id has
been successfully assigned. The absence of this check
stems from the fact [1] that service.py in the ReportPortal
client logs SSL certificate validation failures without
raising exceptions, potentially concealing the issue.

2.
The elimination of the
    os.environ.pop('REQUESTS_CA_BUNDLE', None) workaround
is viable due to the inclusion of the verify_ssl parameter
within the ReportPortalService constructor. This parameter
obviates the need for the aforementioned hack.

[1]
```
        except (ValueError, KeyError, IOError) as exc:
            logger.warning("Report Portal Start Launch request failed",
                           exc_info=exc)
            return
```

Change-Id: I4634661752425f7c282f3e43809b2f1625ee4e40